Moved math code into the new assembly.
