+++ SS14.Client/SS14.Client.csproj
@@ -280,6 +280,10 @@
       <Project>{8af31169-49b1-4a12-b8f4-2a0674a9e7cb}</Project>
       <Name>SS14.Client.Godot</Name>
     </ProjectReference>
+    <ProjectReference Include="..\SS14.Shared.Maths\SS14.Shared.Maths.csproj">
+      <Project>{93f23a82-00c5-4572-964e-e7c9457726d4}</Project>
+      <Name>SS14.Shared.Maths</Name>
+    </ProjectReference>
     <ProjectReference Include="..\SS14.Shared\SS14.Shared.csproj">
       <Project>{0529f740-0000-0000-0000-000000000000}</Project>
       <Name>SS14.Shared</Name>
