+++ SS14.Server/SS14.Server.csproj
@@ -106,6 +106,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\SS14.Shared.Maths\SS14.Shared.Maths.csproj">
+      <Project>{93f23a82-00c5-4572-964e-e7c9457726d4}</Project>
+      <Name>SS14.Shared.Maths</Name>
+    </ProjectReference>
     <ProjectReference Include="..\SS14.Shared\SS14.Shared.csproj">
       <Name>SS14.Shared</Name>
       <Project>{0529F740-0000-0000-0000-000000000000}</Project>
