+++ SS14.Shared.Maths/Circle.cs
@@ -32,8 +32,8 @@ namespace SS14.Shared.Maths
         /// </summary>
         public bool Contains(Vector2 point)
         {
-            var dx = Math.Abs(point.X - Position.X);
-            var dy = Math.Abs(point.Y - Position.Y);
+            var dx = System.Math.Abs(point.X - Position.X);
+            var dy = System.Math.Abs(point.Y - Position.Y);
             var r = Radius;
 
             return (dx * dx + dy * dy <= r * r);
@@ -47,7 +47,7 @@ namespace SS14.Shared.Maths
             var dx = Position.X - circle.Position.X;
             var dy = Position.Y - circle.Position.Y;
 
-            return Math.Sqrt(dx * dx + dy * dy) < Radius + circle.Radius;
+            return System.Math.Sqrt(dx * dx + dy * dy) < Radius + circle.Radius;
         }
 
         /// <summary>
@@ -55,8 +55,8 @@ namespace SS14.Shared.Maths
         /// </summary>
         public bool Intersects(Box2 aabb)
         {
-            var dx = Position.X - Math.Max(aabb.Left, Math.Min(Position.X, aabb.Left + aabb.Height));
-            var dy = Position.Y - Math.Max(aabb.Top, Math.Min(Position.Y, aabb.Top + aabb.Height));
+            var dx = Position.X - System.Math.Max(aabb.Left, System.Math.Min(Position.X, aabb.Left + aabb.Height));
+            var dy = Position.Y - System.Math.Max(aabb.Top, System.Math.Min(Position.Y, aabb.Top + aabb.Height));
             var r = Radius;
 
             return (dx * dx + dy * dy <= r * r);
