+++ SS14.Shared.Maths/Properties/AssemblyInfo.cs
