+++ SS14.Shared.Maths/Ray.cs
@@ -45,7 +45,7 @@ namespace SS14.Shared.Maths
             
             // X axis slab
             {
-                if (Math.Abs(_direction.X) < epsilon)
+                if (System.Math.Abs(_direction.X) < epsilon)
                 {
                     // ray is parallel to this slab, it will never hit unless ray is inside box
                     if (_position.X < FloatMath.Min(box.Left, box.Right) || _position.X > FloatMath.Max(box.Left, box.Right))
@@ -76,7 +76,7 @@ namespace SS14.Shared.Maths
 
             // Y axis slab
             {
-                if (Math.Abs(_direction.Y) < epsilon)
+                if (System.Math.Abs(_direction.Y) < epsilon)
                 {
                     // ray is parallel to this slab, it will never hit unless ray is inside box
                     if (_position.Y < FloatMath.Min(box.Top, box.Bottom) || _position.Y > FloatMath.Max(box.Top, box.Bottom))
